fix: 20207: RecordAccessor / RecordAccessorImpl cleanup
