+++ platform-sdk/swirlds-virtualmap/src/main/java/com/swirlds/virtualmap/internal/reconnect/TeacherPullVirtualTreeView.java
@@ -280,7 +280,7 @@ public final class TeacherPullVirtualTreeView<K extends VirtualKey, V extends Vi
     public void close() {
         try {
             waitUntilReady();
-            records.getDataSource().close();
+            records.close();
         } catch (final IOException e) {
             logger.error(EXCEPTION.getMarker(), "interrupted while attempting to close data source");
         } catch (final InterruptedException e) {
